+++ src/transformers/quantizers/quantizer_gptq.py
@@ -1,0 +1,91 @@
+# Copyright 2024 The HuggingFace Inc. team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from .base import HFQuantizer
+from packaging import version
+import importlib
+
+from typing import TYPE_CHECKING, Any, Dict, List, Union, Optional
+
+if TYPE_CHECKING:
+    from ..modeling_utils import PreTrainedModel
+
+from ..utils import is_torch_available, logging, is_optimum_available, is_auto_gptq_available
+from ..utils.quantization_config import QuantizationConfigMixin, GPTQConfig
+
+if is_torch_available():
+    import torch
+
+logger = logging.get_logger(__name__)
+
+class GptqHFQuantizer(HFQuantizer):
+    """
+    Quantizer of the GPTQ method - for GPTQ the quantizer support calibration of the model through
+    `auto_gptq` package. Quantization is done under the hood for users if they load a non-prequantized model.
+    """
+    # TODO: what is this?
+    requires_parameters_quantization = False
+
+    requires_calibration = False
+    required_packages = ["optimum", "auto_gptq"]
+
+    def __init__(self, quantization_config: QuantizationConfigMixin, **kwargs):
+        super().__init__(quantization_config, **kwargs)
+        from optimum.gptq import GPTQQuantizer
+
+        self.quantizer = GPTQQuantizer.from_dict(self.quantization_config.to_dict())
+
+    def validate_environment(self, *args, **kwargs):
+        gptq_supports_cpu = version.parse(importlib.metadata.version("auto-gptq")) > version.parse("0.4.2")
+        if not gptq_supports_cpu and not torch.cuda.is_available():
+            raise RuntimeError("GPU is required to quantize or run quantize model.")
+        elif not (is_optimum_available() and is_auto_gptq_available()):
+            raise ImportError(
+                "Loading a GPTQ quantized model requires optimum (`pip install optimum`) and auto-gptq library (`pip install auto-gptq`)"
+            )
+        elif version.parse(importlib.metadata.version("auto_gptq")) < version.parse("0.4.2"):
+            raise ImportError(
+                "You need a version of auto_gptq >= 0.4.2 to use GPTQ: `pip install --upgrade auto-gptq`"
+            )
+
+    def set_torch_dtype(self, torch_dtype: torch.dtype) -> torch.dtype:
+        if torch_dtype is None:
+            torch_dtype = torch.float16
+        elif torch_dtype != torch.float16:
+            logger.info("We suggest you to set `torch_dtype=torch.float16` for better efficiency with GPTQ.")
+        return torch_dtype
+
+    def _process_model_before_weight_loading(self, model: "PreTrainedModel", **kwargs):
+        if model.__class__.main_input_name != "input_ids":
+            raise RuntimeError("We can only quantize pure text model.")
+
+        if self.pre_quantized:
+            model = self.quantizer.convert_model(model)
+
+    def _process_model_after_weight_loading(self, model: "PreTrainedModel", **kwargs):
+        if self.pre_quantized:
+            model = self.quantizer.post_init_model(model)
+        else:
+            if self.quantization_config.tokenizer is None:
+                self.quantization_config.tokenizer = model.name_or_path
+
+            self.quantizer.quantize_model(model, self.quantization_config.tokenizer)
+            model.config.quantization_config = GPTQConfig.from_dict(self.quantizer.to_dict())
+
+    @property
+    def is_trainable(self, model: Optional["PreTrainedModel"] = None):
+        return True
+
+    @property
+    def is_serializable(self):
+        return True
